+++ src/components/Accounts/ChargeFeeFine/ChargeFeeFine.js
@@ -10,7 +10,6 @@ import {
 } from '@folio/stripes/components';
 import {
   FormattedMessage,
-  intlShape,
   injectIntl
 } from 'react-intl';
 import SafeHTMLMessage from '@folio/react-intl-safe-html';
@@ -51,7 +50,7 @@ class ChargeFeeFine extends React.Component {
     okapi: PropTypes.object,
     selectedLoan: PropTypes.object,
     user: PropTypes.object,
-    intl: intlShape.isRequired,
+    intl: PropTypes.object.isRequired,
     history: PropTypes.object,
     location: PropTypes.object,
     match: PropTypes.object,
@@ -121,7 +120,7 @@ class ChargeFeeFine extends React.Component {
     type.title = item.title;
     type.barcode = item.barcode;
     type.callNumber = item.callNumber;
-    type.location = (selectedLoan.id) ? (item.location || {}).name : (item.effectiveLocation || {}).name;
+    type.location = item?.location?.name || item?.effectiveLocation?.name;
     type.materialType = (item.materialType || {}).name;
     type.materialTypeId = (selectedLoan.id) ? '0' : (item.materialType || {}).id || '0';