+++ src/components/Loans/ClosedLoans/ClosedLoans.js
@@ -4,7 +4,6 @@ import {
   FormattedMessage,
   FormattedTime,
   injectIntl,
-  intlShape,
 } from 'react-intl';
 import PropTypes from 'prop-types';
 
@@ -22,7 +21,7 @@ import {
   IntlConsumer,
   stripesShape,
 } from '@folio/stripes/core';
-import { effectiveCallNumber } from '@folio/stripes-util';
+import { effectiveCallNumber } from '@folio/stripes/util';
 
 import {
   calculateSortParams,
@@ -49,7 +48,7 @@ class ClosedLoans extends React.Component {
     }),
     match: PropTypes.object,
     user: PropTypes.object,
-    intl: intlShape.isRequired,
+    intl: PropTypes.object.isRequired,
     history: PropTypes.object.isRequired,
     location: PropTypes.object.isRequired,
     handleOptionsChange: PropTypes.func.isRequired,
@@ -306,10 +305,8 @@ class ClosedLoans extends React.Component {
 
   buildRecords(records) {
     return records.map((record) => {
-      const {
-        item,
-        item: { contributors },
-      } = record;
+      const { item } = record;
+      const contributors = item?.contributors;
 
       return _.isArray(contributors) ?
         {
@@ -340,19 +337,21 @@ class ClosedLoans extends React.Component {
 
     const accounts = _.get(this.props.resources, ['loanAccount', 'records'], []);
     const accountsLoan = accounts.filter(a => a.loanId === loan.id) || [];
-    const itemDetailsLink = `/inventory/view/${loan.item.instanceId}/${loan.item.holdingsRecordId}/${loan.itemId}`;
+    const itemDetailsLink = loan.item && `/inventory/view/${loan.item.instanceId}/${loan.item.holdingsRecordId}/${loan.itemId}`;
     const buttonDisabled = !stripes.hasPerm('ui-users.feesfines.actions.all');
 
     return (
       <DropdownMenu data-role="menu">
-        <IfPermission perm="inventory.items.item.get">
-          <Button
-            buttonStyle="dropdownItem"
-            to={itemDetailsLink}
-          >
-            <FormattedMessage id="ui-users.itemDetails" />
-          </Button>
-        </IfPermission>
+        {itemDetailsLink &&
+          <IfPermission perm="inventory.items.item.get">
+            <Button
+              buttonStyle="dropdownItem"
+              to={itemDetailsLink}
+            >
+              <FormattedMessage id="ui-users.itemDetails" />
+            </Button>
+          </IfPermission>
+        }
         <Button
           disabled={buttonDisabled}
           buttonStyle="dropdownItem"