+++ src/views/LoanDetails/LoanDetails.js
@@ -3,7 +3,6 @@ import {
   FormattedMessage,
   FormattedTime,
   injectIntl,
-  intlShape,
 } from 'react-intl';
 import { compose } from 'redux';
 import Link from 'react-router-dom/Link';
@@ -28,7 +27,7 @@ import {
   NoValue,
 } from '@folio/stripes/components';
 import { IfPermission } from '@folio/stripes/core';
-import { effectiveCallNumber } from '@folio/stripes-util';
+import { effectiveCallNumber } from '@folio/stripes/util';
 
 import PatronBlockModal from '../../components/PatronBlock/PatronBlockModal';
 import {
@@ -81,7 +80,7 @@ class LoanDetails extends React.Component {
     declareLost: PropTypes.func,
     claimReturned: PropTypes.func,
     patronBlocks: PropTypes.arrayOf(PropTypes.object),
-    intl: intlShape.isRequired,
+    intl: PropTypes.object.isRequired,
     match: PropTypes.object.isRequired,
     history: PropTypes.object.isRequired,
     location: PropTypes.object.isRequired,
@@ -389,14 +388,16 @@ class LoanDetails extends React.Component {
                     <FormattedMessage id="ui-users.renew" />
                   </Button>
                 </IfPermission>
-                <Button
-                  data-test-claim-returned-button
-                  disabled={buttonDisabled || itemStatus === 'Claimed returned'}
-                  buttonStyle="primary"
-                  onClick={() => claimReturned(loan)}
-                >
-                  <FormattedMessage id="ui-users.loans.claimReturned" />
-                </Button>
+                <IfPermission perm="ui-users.loans.claim-item-returned">
+                  <Button
+                    data-test-claim-returned-button
+                    disabled={buttonDisabled || itemStatus === 'Claimed returned'}
+                    buttonStyle="primary"
+                    onClick={() => claimReturned(loan)}
+                  >
+                    <FormattedMessage id="ui-users.loans.claimReturned" />
+                  </Button>
+                </IfPermission>
                 <IfPermission perm="ui-users.loans.edit">
                   <Button
                     disabled={buttonDisabled || isDeclaredLostItem}